requirements.txt
# Core dependencies
pydantic>=2.0.0
dataclasses-json>=0.6.0
python-dotenv>=1.0.0

# Message queue support
pika>=1.3.0  # RabbitMQ
aiokafka>=0.10.0  # Kafka
redis>=5.0.0  # Redis pub/sub
kombu>=5.3.0  # Message queue abstraction

# MCP and Composio integration
composio-core>=0.5.0
composio-langchain>=0.5.0
mcp>=0.1.0

# GitHub integration
PyGithub>=2.1.0
gitpython>=3.1.0

# Code analysis and AST
astroid>=3.0.0
ast-grep-py>=0.12.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
rope>=1.11.0  # Refactoring library

# Testing frameworks
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
coverage>=7.3.0
hypothesis>=6.90.0

# Code formatting and linting
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
isort>=5.12.0
autopep8>=2.0.0

# Python framework support
django>=4.2.0
fastapi>=0.104.0
flask>=3.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0

# Async and concurrency
aiohttp>=3.9.0
aiofiles>=23.0.0
asyncio>=3.4.3

# Utilities
pyyaml>=6.0.0
toml>=0.10.0
click>=8.1.0
rich>=13.0.0  # Better console output
tenacity>=8.2.0  # Retry logic

# Development dependencies
ipython>=8.17.0
jupyter>=1.0.0
python-lsp-server>=1.9.0